Revert "Pbi4 subtask4.2"
