+++ .gitignore
@@ -44,5 +44,3 @@ next-env.d.ts
 .sentryclirc
 
 /.vscode
-
-/.sonarlint
