+++ jest-setup.ts
@@ -1,19 +1,4 @@
 import '@testing-library/jest-dom';
 import React from 'react';
 
-Object.defineProperty(window, 'matchMedia', {
-    writable: true,
-    value: jest.fn().mockImplementation((query) => ({
-      matches: false,
-      media: query,
-      onchange: null,
-      addListener: jest.fn(),
-      removeListener: jest.fn(),
-      addEventListener: jest.fn(),
-      removeEventListener: jest.fn(),
-      dispatchEvent: jest.fn(),
-    })),
-});
-
 global.React = React;
-global.console.log = jest.fn();
