+++ sonar-project.properties
@@ -1,9 +1,9 @@
-sonar.projectKey=MAAMS-NG-FE
-sonar.host.url=https://sonarqube.cs.ui.ac.id
+sonar.projectKey=Kelompok-5-PPL-A_MAAMS-NG-FE
+sonar.organization=kelompok-5-ppl-a-1
+sonar.host.url=https://sonarcloud.io
 sonar.login=${SONAR_TOKEN}
 
 sonar.sources=src
 sonar.tests=src/__tests__
 sonar.test.inclusions=**/*.test.tsx
-
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
