--- src/__tests__/constants/mode.test.tsx
@@ -1,8 +1,0 @@
-import Mode from '../../constants/mode';
-
-describe('Mode Enum', () => {
-  test('should have correct values', () => {
-    expect(Mode.pribadi).toBe('PRIBADI');
-    expect(Mode.pengawasan).toBe('PENGAWASAN');
-  });
-});
