--- src/components/types/cause.d.ts
@@ -1,13 +1,0 @@
-export interface Cause {
-    root_status: boolean
-    id: string
-    problem: string
-    row: number
-    column: number
-    mode: string
-    cause: string
-    status: boolean
-    feedback: string
-    root_status: boolean
-  }
-
