--- src/components/types/cell.d.ts
@@ -1,10 +1,0 @@
-export interface CellProps {
-    cellName: string
-    cause: string
-    onChange: (value: string) => void
-    causeStatus: CauseStatus
-    disabled: boolean
-    placeholder: string
-    feedback?: string
-  }
-
