--- src/components/types/circularIconButtonProps.d.ts
@@ -1,8 +1,0 @@
-import { ReactNode } from 'react'
-
-export interface CircularIconButtonProps {
-  icon: ReactNode
-  onClick?: MouseEventHandler<HTMLButtonElement>
-  type: 'button' | 'submit' | 'reset'
-  id?: string
-}
