--- src/components/types/counterButton.d.ts
@@ -1,6 +1,0 @@
-export interface CounterButtonProps {
-    number: number
-    onIncrement: () => void
-    onDecrement: () => void
-  }
-
