--- src/components/types/deleteButtonProps.d.ts
@@ -1,5 +1,0 @@
-export interface DeleteButtonProps {
-    idQuestion: string | string[] | undefined
-    pathname: string
-  }
-
