--- src/components/types/editableTitleForm.d.ts
@@ -1,6 +1,0 @@
-interface EditableTitleFormProps {
-    title: string | undefined
-    id: string | string[] | undefined
-    onTitleChange: (newTitle: string) => void
-  }
-
