--- src/components/types/row.d.ts
@@ -1,11 +1,0 @@
-export interface RowProps {
-    rowNumber: number
-    cols: number
-    causes: string[]
-    causesId: string[]
-    causeStatuses: CauseStatus[]
-    disabledCells: boolean[]
-    onCauseAndStatusChanges: (causeIndex: number, newValue: string, newStatus: CauseStatus) => void
-    feedbacks: string[]
-  }
-
