--- src/lib/enum.ts
@@ -1,8 +1,0 @@
-export enum CauseStatus {
-    Incorrect = 'INCORRECT',
-    CorrectNotRoot = 'CORRECT_NOT_ROOT',
-    CorrectRoot = 'CORRECT_ROOT',
-    Resolved = 'RESOLVED',
-    Unchecked = 'UNCHECKED'
-  }
-
