+++ src/pages/_app.tsx
@@ -1,12 +1,10 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 import { ChakraProvider } from "@chakra-ui/react";
-import { Toaster } from "react-hot-toast";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <ChakraProvider>
-      <Toaster />
       <Component {...pageProps} />;
     </ChakraProvider>
   )
