+++ src/pages/validator/index.tsx
@@ -7,7 +7,6 @@ import { useRouter } from 'next/router'
 import { CustomInput } from '@/components/customInput'
 import { Badge } from '@/badges'
 import ConfirmationPopup from '@/components/confirmationPopup'
-import axiosInstance from '../../services/axiosInstance'
 
 const QuestionAddPage: React.FC = () => {
   const router = useRouter()
@@ -16,7 +15,7 @@ const QuestionAddPage: React.FC = () => {
   const [title, setTitle] = useState<string>('')
   const [question, setQuestion] = useState<string>('')
   const [newTag, setNewTag] = useState<string>('')
-  const [isLoading, setIsLoading] = useState<boolean>(false)
+  const [isLoading] = useState<boolean>(false)
   const [tags, setTags] = useState<string[]>([])
   const [showConfirmation, setShowConfirmation] = useState<boolean>(false)
   const [selectedMode, setSelectedMode] = useState<Mode>(mode)
@@ -60,46 +59,6 @@ const QuestionAddPage: React.FC = () => {
       }
       setTags((prevCategories) => [...prevCategories, newTag.trim()])
       setNewTag('')
-    }
-  }
-  
-  const handleSubmit = async () => {
-    setIsLoading(true)
-    if (!title) {
-      toast.error('Judul harus diisi')
-      setIsLoading(false)
-      return
-    } else if (!question) {
-      toast.error('Pertanyaan harus diisi')
-      setIsLoading(false)
-      return
-    } else if (tags.length == 0) {
-      toast.error('Minimal mengisi 1 kategori')
-      setIsLoading(false)
-      return
-    } else if (title.length > 40) {
-      toast.error('Judul maksimal 40 karakter. Berikan judul yang lebih singkat')
-      setIsLoading(false)
-      return
-    }
-
-    try {
-      const { data } = await axiosInstance.post('/question/submit/', {
-        title: title,
-        question: question,
-        mode: mode,
-        tags: tags
-      })
-      toast.success('Analisis berhasil ditambahkan')
-      router.push(`/validator/${data.id}`)
-    } catch (error: any) {
-      if (error.response) {
-        toast.error(error.response.data.detail)
-        setIsLoading(false)
-      } else {
-        toast.error('Gagal menambahkan analisis')
-        setIsLoading(false)
-      }
     }
   }
 
@@ -158,8 +117,8 @@ const QuestionAddPage: React.FC = () => {
             <div className='flex justify-center w-full flex-col lg:flex-row'>
               <button
                 type='button'
-                onClick={handleSubmit}
-                className='bg-gradient-to-b from-yellow-400 to-yellow-600 text-l text-white font-bold py-2 px-12 rounded-xl'
+                // onClick={handleSubmit}
+                className='bg-gradient-to-b from-yellow-400 to-yellow-600 text-l text-white font-bold py-2 px-12 rounded-xl w-[192px] mx-auto'
                 disabled={isLoading}
               >
                 Kirim
