--- src/services/axiosInstance.ts
@@ -1,7 +1,0 @@
-import axios from 'axios'
-
-const axiosInstance = axios.create({
-  baseURL: `${process.env.NEXT_PUBLIC_API_BASE_URL}`
-})
-
-export default axiosInstance
